# docs/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DeNoise application API",
    "description": "Authenticated operator API (docs/06-api-contract.md). Session cookie for the UI or \u0060Authorization: Bearer ah_pat_\u003CkeyId\u003E.\u003Csecret\u003E\u0060 for API clients.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/teams": {
      "get": {
        "tags": [
          "Teams"
        ],
        "operationId": "ListTeams",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamSummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "post": {
        "tags": [
          "Teams"
        ],
        "operationId": "CreateTeam",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamSummary"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/teams/{id}": {
      "get": {
        "tags": [
          "Teams"
        ],
        "operationId": "GetTeam",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamSummary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/teams/{id}/overview": {
      "get": {
        "tags": [
          "Teams"
        ],
        "operationId": "GetTeamOverview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamOverview"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/teams/{id}/members/{userId}": {
      "post": {
        "tags": [
          "Teams"
        ],
        "operationId": "AddTeamMember",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Teams"
        ],
        "operationId": "RemoveTeamMember",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
    "/api/v1/integrations": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "operationId": "ListIntegrations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegrationSummary"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "post": {
        "tags": [
          "Integrations"
        ],
        "operationId": "CreateIntegration",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIntegrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
...
    "/api/v1/integrations/{id}/divergence": {
      "get": {
        "tags": [
          "Integrations"
        ],
        "operationId": "GetIntegrationDivergence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DivergenceReportDto"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      },
      "post": {
        "tags": [
          "Integrations"
        ],
        "operationId": "RunIntegrationDivergence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DivergenceReportDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
...
    "/api/v1/integrations/{id}/rotate-ingest-token": {
      "post": {
        "tags": [
          "Integrations"
        ],
        "operationId": "RotateIngestToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationCreatedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "session": []
          },
          {
            "pat": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTeamRequest": {
        "required": [
          "name",
          "accessScopes"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "accessScopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isTriage": {
            "type": "boolean",
            "default": false
          },
          "fallbackTeamId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "defaultEscalationPolicyId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "description": "Server-side session cookie set by POST /auth/login.",
        "name": "denoise_session",
        "in": "cookie"
      },
      "pat": {
        "type": "http",
        "description": "Personal access token ah_pat_\u003CkeyId\u003E.\u003Csecret\u003E.",
        "scheme": "bearer"
      }
    }
  },
...
}